tweaked dyn config, fixed bug wth RBGColor type, added saturation bands, added costCloud and costNormCloud pubs, commented out experimental region-growing method, fixed bug with off-normal cost calc, updated launchfiles

+++ src/stair_detection_node.cpp
@@ -52,7 +52,7 @@ ros::Publisher seg_stairs_pub_;
 ros::Publisher stair_parts_pub_;
 ros::Publisher nearest_step_pose_pub_;
 
-struct RGBColor{ uint r, g, b; };
+struct RGBColor{ float r, g, b; };
 inline RGBColor getRGBColor(double ratio)
 {
     //we want to normalize ratio so that it fits in to 6 regions
